dev-requirements.txt
-r requirements.txt

nose>=1.3.0
unittest2>=1.1.0
mock>=1.3.0
coverage>=3.7.1
behave>=1.2.5
docker-py>=1.10.3
